Allow using Koji as a source of information
We might want to install Flatpaks via an index before they are pushed
publically. The source of truth for what is going to go out in a future
release is a brew tag - so allow configuring an index to index all Flatpaks
with a particular brew/koji tag. The index should be pointed to a
registry configuration that a) has an appropriate registry URL configured
matching the pullspecs in brew b) has the correct koji config name configured.