Add temporary workaround for an atomic-reactor bug
The metainfo that atomic-reactor is creating for Flatpaks that
are converted from OCI to docker format on upload has an id
in the archive metadata that matches the OCI image, not the
docker image, which confuses Pyxis. Work around that for
koji-backed repositories by digging out the actual digest we
need and using that to override the information from Pyxis.