Build tar-diff binary into the indexer container
Add https://github.com/containers/tardiff as a submodule, and make
the indexer container build a two stage process: first build an
image with the tar-diff container, then use generated tar-diff
binary as an input to the s2i build of the indexer image.

This should be compatible with the "image source" capability of
OpenShift builds.