tests/utils: Add WithArgDecorator, use for mock_fedora_messaging
WithArgDecorator is a class used to create a decorator that that does
some setup/teardown around a test case, and optionally passes a named argument
into the test case.

Example:
   mock_fedora_messaging = WithArgDecorator('mock_connection', _setup_fedora_messaging)

If a test function decoratated with @mock_fedora_messaging has a mock_connection
parameter, then the value that the _setup_fedora_messaging generator yields will
be passed in as that paramater.