Add delta generation to indexing
If 'delta_keep_days' is set in an index config, then images are considered
candidates for delta updates until the next newer image in the tag history
at least delta_keep_days days old.

Add work items to Redis to generate layer differences between such candidates
and the latest image for the tag, and collect the results and write a
delta manifest for the latest image.

The implementation of running tar-diff to create the layer differences will
be in a separate commit.