Add a 'differ' mode to process work tar-diff work items
Add a second daemon mode to the cli 'flatpak_indexer differ' - that waits for
work items to be added to the tardiff:pending queue and processes them by
running tar-diff.

The idea here is that running tar-diff can be parallelized across multiple
containers - so that if there are many Flatpaks to update - and if there
are multiple architectures for Flatpak - we can get the deltas done quickly.