pyxis_updater: rework to support deltas
Write an appropriate TagHistory into the data that is saved for the indexer
to allow generating deltas. Switch to using the koji_query code that was
written for the Flatpak updater instead of separate code to query brew, and
rely on this more heavily: Pyxis is used only to get the tag history and
current tags, and not to get immmutable image information that is also available
through brew.