Remove work directory and writing of intermediate files
The intermediate files were added from an idea that the updater and the
indexer might be separate processes - but that doesn't seem necessary
at this point. Simplify things by keeping the intermediate data in memory.

Short-circuiting is preserved for subsequent runs of the indexer, but
is never done on the first run - making sure that we get an up-to-date
index if the code or config changes.