tests/bodhi.py: strip out update builds that aren't in the test data
Including all builds for all updates that we reference would create huge test
data, so we only include a partial set of builds in the test data. Avoid problems
by stripping the builds that we don't include in the test data out of the
updates.

This required adding some more RPM builds to fix a test where which looked at
RPM updates build nvrs without looking at the builds.