Add support for indexes with Flatpak metadata in annotations
Setting 'flatpak_annotations' in the indexer config will cause the
index to be written with the org.flatpak and org.freedesktop annotations
labels moved to annotations. This is to allow for making a server config
that is compatible with both flatpak < 1.6 (which expects metadata in
annotations) and flatpak >= 1.6 (which expects metadata in labels.)
The indexed images need to have both labels and annotations for this
to work.