Cleaner: check for and remove dangling tardiff:result objects
When we delete a tardiff off of disk because it is no-longer referenced,
leaving the tardiff:result object around is harmful: we'll assume that
it exists and not try to regenerate it.

In the clean phase, scan all tardiff:result objects and make sure that
they still exist on disk - if not, remove the result object.