bodhi_change_monitor: handle exceptions when talking to fedora-messaging
If we get a StreamLostError exception, try to reconnect to fedora-messaging,
with exponential backoff.

if we get any other exception, rethrow that in the  main thread, the next
time monitor.get_changed() is called. This isn't ideal, but seems better
than just exiting the process or adding a event loop (asyncio or twisted)
to the main process so we can rethrow sooner.

See https://pagure.io/fedora-infrastructure/issue/9794