Add some fallback handle if we encounter a module with multiple contexts
If a user builds the same git commit of a module with different platform buildrequires
(8.4.0 and 8.4.0-z, for example), then we can end up with multiple contexts.
The guess here corresponds to what ODCS will do (use the most recent), but the
better thing would be to avoid this situation and always use the full NSVC to
look up a build.