Add ability to create a content index
To be able to figure out automated rebuilds, we might want to be able to
answer the question: "what flatpaks that we are distributing are built
from flatpak-common:f33".

A new per-index config option "contents:" is added that specifies a
contents index directory, with files such as:

 <contents_directory>/modules/flatpak-common:f33.json

that specify what images are built from that module, and what packages
from the module the image uses.