Redo handling of local_certs to be redirect friendly
The registry we use to get images for Red Hat now redirects to quay.io, so
pinning the requests.get() to expect a local certificate didn't work
properly. Instead of passing the certificate into the requests objects,
use a HTTPAdapter subclass that knows how to look up SSL certificates in
our config.