bodhi_change_monitor: Handle reconnection to a different queue
When we reconnect to fedora-messaging, we might get a different queue,
and thus have no idea about what has changed - handle this by returning
a tuple of queue_name, <changes> from the change monitor. If we get
a different queue name when we go to update, we need reset everything
at that point.