bodhi_change_monitor: Handle race conditions with monitor.stop()
With reconnection, monitor.stop() could be called at points where monitor.connection
pointed to an old connection instead of the connection that would next be
iterated. Make sure that monitor.stop() reliably works at any point.

This required redoing the fedora_messaging mock code so that we aren't adding
events to the current connection, but events that could be consumed
from *any* connection.