json_model: add simple schema migration support
Add a check_json_current() vfunc to BaseModel - this allows implementing
a check that the JSON corresponds to the current version of the model's
schema. And add a check_current flag to BaseModel.from_json() /
from_json_text() to have them call check_json_current() and return
None on failure. This allows dropping cached items and refetching them.