When possible, use ODCS to resolve source modules more accurately
The metadata for Flatpak builds contains a modules key that lists the
name-stream-version for each module used. Unfortunately, it doesn't
contain the context. In the simple case, there's only one module build
for the name-stream-version. But sometimes that doesn't work and we
don't know what to pick. We could:

 Module was built
 Module was rebuilt *without a new commit*
 Image was built

 Image was built
 Module was rebuilt *without a new commit*

Newer versions of OSBS have odcs: compose_ids in the metadata - when
that is available, we can ask ODCS for information about the composes
and use that to find the full name-stream-version.context without
guessing.