Switch to using Pyxis's graphql api, drop use of tag history
Use the more modern graphql API to Pyxis rather than the
legacy REST API. This isn't absolutely necessary, but simplifies
things and makes our queries more transparent in what they
are doing. It also probably is useful future-proofing.

This adds a test dependency on graphql-core, so that we can parse
the graphql queries and execute them (in super-simplified fashion)
against the test data.

Stopping using the TagHistory information from Konflux

 - The tag history from Metaxor has been problematical for years -
   when it's not correct, we index the wrong images
 - Konflux does not write TagHistory data at the moment

As a replacement, we just look at the push_date of containers in
the repository and assume that everything in the repository was
tagged with the target tag at the date it was pushed - this
really only works for "latest", but we care most about that.