Cache image data when queried from the registry
The Pyxis updater now queries the manifest/config for any image not in
Koji from the registry. To increase performance and robustness, cache
the image data we generate from the registry data in Redis.

The cached ImageModel omits certain data that is either variable
(tags) or depends on our config (the pullspec).