Work around deficiencies in redis-py typing
redis-py has problematical type annotations - async and sync return types
are mixed together, and there is no representation of the fact that
the client *might* decode the responses from bytes to str, or not,
depending on how it was constructed.

See https://github.com/redis/redis-py/pull/3619

Work around this by adding protocols that represent the types of the
API as we use it, and cast to the protocols as needed.