Fix FLATPAK_INDEXER_UPDATE_TEST_DATA for openshift
When building in openshift with the Docker strategy, the
filesystem ownership mismatches the UID causing git to refuse
fetch test data. Configure git to treat the source directory
as safe.

Also fix mistake in the containerfile - FLATPAK_INDEXER_UPDATE_TEST_DATA
has an effect if it's set to any non-null value, not just "true",
so use "" as the default value.